+++ backend-lib/elykia-client/src/main/java/com/optimize/elykia/client/dto/ClientRespDto.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.optimize.elykia.client.entity.Client;
 import com.optimize.elykia.client.enumeration.ClientType;
-import lombok.ToString;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -17,8 +16,9 @@ public record ClientRespDto(Long id, String firstname, String lastname, String a
                             String businessCreditAuthorizedBy, LocalDateTime businessCreditAuthorizedAt,
                             String occupation, ClientType clientType, Double latitude,
                             Double longitude, String mll, LocalDate syncDate, String code, String profilPhotoUrl,
-                            String cardPhotoUrl, String tontineCollector, @JsonFormat(pattern = "yyyy-MM-dd") LocalDateTime createdAt
-
+                            String cardPhotoUrl, String tontineCollector,
+                            @JsonFormat(pattern = "yyyy-MM-dd") LocalDateTime createdAt,
+                            String profilPhotoThumbUrl, String cardPhotoThumbUrl
 ) {
 
     @JsonIgnore
@@ -26,7 +26,7 @@ public record ClientRespDto(Long id, String firstname, String lastname, String a
         return new ClientRespDto(id, null, null, null, null, null, null,
                 null, null, null, null, null,
                 null, null, null, null, null, null, null,
-                null, null, null, null, null, null, null, null, null, null);
+                null, null, null, null, null, null, null, null, null, null, null, null);
     }
 
     @JsonIgnore
@@ -34,20 +34,22 @@ public record ClientRespDto(Long id, String firstname, String lastname, String a
         return new ClientRespDto(id, firstname, lastname, null, null, null, null,
                 null, null, null, null, null,
                 null, null, null, null, null, null, null,
-                null, null, null, null, null, null, null, null, null, null);
+                null, null, null, null, null, null, null, null, null, null, null, null);
     }
-
 
     @JsonIgnore
     public static ClientRespDto fromClient(Client client) {
         if (Objects.isNull(client)) {
-            return  null;
+            return null;
         }
-        return new ClientRespDto(client.getId(), client.getFirstname(), client.getLastname(), client.getAddress(), client.getPhone(), client.getCardID(), client.getCardType(),
+        return new ClientRespDto(client.getId(), client.getFirstname(), client.getLastname(), client.getAddress(),
+                client.getPhone(), client.getCardID(), client.getCardType(),
                 client.getDateOfBirth(), null, null, null, client.getCollector(),
                 client.getQuarter(), client.getCreditInProgress(), client.isBusinessCreditInProgress(),
                 client.isBusinessCreditAuthorized(), client.getBusinessCreditAuthorizedBy(),
-                client.getBusinessCreditAuthorizedAt(), client.getOccupation(), client.getClientType(), null, null, null,
-                null, client.getCode(), client.getProfilPhotoUrl(), client.getCardPhotoUrl(), client.getTontineCollector(), client.getCreatedDate());
+                client.getBusinessCreditAuthorizedAt(), client.getOccupation(), client.getClientType(),
+                null, null, null, null, client.getCode(), client.getProfilPhotoUrl(), client.getCardPhotoUrl(),
+                client.getTontineCollector(), client.getCreatedDate(),
+                client.getProfilPhotoThumbUrl(), client.getCardPhotoThumbUrl());
     }
 }