+++ backend-lib/elykia-client/src/main/java/com/optimize/elykia/client/repository/ClientRepository.java
@@ -66,7 +66,7 @@ public interface ClientRepository extends GenericRepository<Client, Long> {
                 c.firstname, c.lastname, c.address, c.phone, c.cardID, c.cardType, c.dateOfBirth,
                 c.contactPersonName, c.contactPersonPhone, c.contactPersonAddress, c.collector,
                 c.quarter, c.creditInProgress, c.businessCreditInProgress, c.businessCreditAuthorized, c.businessCreditAuthorizedBy, c.businessCreditAuthorizedAt, c.occupation, c.clientType, c.latitude, c.longitude,
-                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate)
+                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate, c.profilPhotoThumbUrl, c.cardPhotoThumbUrl)
                 FROM Client c
                 WHERE (c.collector = :collector OR c.tontineCollector = :collector OR c.recoveryCollector = :collector) AND c.clientType = :clientType AND c.state = :state
             """)
@@ -78,7 +78,7 @@ public interface ClientRepository extends GenericRepository<Client, Long> {
                 c.firstname, c.lastname, c.address, c.phone, c.cardID, c.cardType, c.dateOfBirth,
                 c.contactPersonName, c.contactPersonPhone, c.contactPersonAddress, c.tontineCollector,
                 c.quarter, c.creditInProgress, c.businessCreditInProgress, c.businessCreditAuthorized, c.businessCreditAuthorizedBy, c.businessCreditAuthorizedAt, c.occupation, c.clientType, c.latitude, c.longitude,
-                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate)
+                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate, c.profilPhotoThumbUrl, c.cardPhotoThumbUrl)
                 FROM Client c
                 WHERE c.tontineCollector = :collector AND c.clientType = :clientType AND c.state = :state
             """)
@@ -90,7 +90,7 @@ public interface ClientRepository extends GenericRepository<Client, Long> {
                 c.firstname, c.lastname, c.address, c.phone, c.cardID, c.cardType, c.dateOfBirth,
                 c.contactPersonName, c.contactPersonPhone, c.contactPersonAddress, c.collector,
                 c.quarter, c.creditInProgress, c.businessCreditInProgress, c.businessCreditAuthorized, c.businessCreditAuthorizedBy, c.businessCreditAuthorizedAt, c.occupation, c.clientType, c.latitude, c.longitude,
-                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate)
+                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate, c.profilPhotoThumbUrl, c.cardPhotoThumbUrl)
                 FROM Client c
                 WHERE (c.tontineCollector = :collector OR c.collector = :collector OR c.recoveryCollector = :collector) AND c.clientType = :clientType AND c.state = :state
             """)
@@ -102,7 +102,7 @@ public interface ClientRepository extends GenericRepository<Client, Long> {
                 c.firstname, c.lastname, c.address, c.phone, c.cardID, c.cardType, c.dateOfBirth,
                 c.contactPersonName, c.contactPersonPhone, c.contactPersonAddress, c.collector,
                 c.quarter, c.creditInProgress, c.businessCreditInProgress, c.businessCreditAuthorized, c.businessCreditAuthorizedBy, c.businessCreditAuthorizedAt, c.occupation, c.clientType, c.latitude, c.longitude,
-                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate)
+                c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate, c.profilPhotoThumbUrl, c.cardPhotoThumbUrl)
                 FROM Client c
                 WHERE c.state <> :state
 
@@ -147,7 +147,7 @@ public interface ClientRepository extends GenericRepository<Client, Long> {
 
     Optional<Client> findFirstByPhoneAndClientTypeAndState(String phone, ClientType clientType, State state);
 
-    @Query("SELECT new com.optimize.elykia.client.dto.ClientRespDto(c.id, c.firstname, c.lastname, c.address, c.phone, c.cardID, c.cardType, c.dateOfBirth, c.contactPersonName, c.contactPersonPhone, c.contactPersonAddress, c.collector, c.quarter, c.creditInProgress, c.businessCreditInProgress, c.businessCreditAuthorized, c.businessCreditAuthorizedBy, c.businessCreditAuthorizedAt, c.occupation, c.clientType, c.latitude, c.longitude, c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate) " +
+    @Query("SELECT new com.optimize.elykia.client.dto.ClientRespDto(c.id, c.firstname, c.lastname, c.address, c.phone, c.cardID, c.cardType, c.dateOfBirth, c.contactPersonName, c.contactPersonPhone, c.contactPersonAddress, c.collector, c.quarter, c.creditInProgress, c.businessCreditInProgress, c.businessCreditAuthorized, c.businessCreditAuthorizedBy, c.businessCreditAuthorizedAt, c.occupation, c.clientType, c.latitude, c.longitude, c.mll, c.syncDate, c.code, c.profilPhotoUrl, c.cardPhotoUrl, c.tontineCollector, c.createdDate, c.profilPhotoThumbUrl, c.cardPhotoThumbUrl) " +
        "FROM Client c " +
        "WHERE c.state <> com.optimize.common.entities.enums.State.DELETED " +
        "AND (:#{#username == null} = true OR ( " +